Implement notification feature

+++ src/stores/taskStore.ts
@@ -3,6 +3,7 @@ import { ref } from 'vue'
 import type { Task, TaskFormData } from '@/types/Task.ts'
 import { useRouter } from 'vue-router'
 import useTask from '@/composables/services/TaskService.ts'
+import useNotification from '@/composables/notification.ts'
 export const useTaskStore = defineStore('taskStore', () => {
   const tasks = ref<Task[]>([])
   const errors = ref({})
@@ -13,8 +14,10 @@ export const useTaskStore = defineStore('taskStore', () => {
     description: '',
     task_status_id: null,
   })
+
   const router = useRouter()
   const { getTasks, createTask, getTaskById, updateTaskById, deleteTaskById } = useTask()
+  const { confirm,success, error} = useNotification()
 
   const resetForm = () => {
     form.value.title = ''
@@ -29,8 +32,8 @@ export const useTaskStore = defineStore('taskStore', () => {
     loading.value = true
     try {
       tasks.value = await getTasks()
-    } catch (error) {
-      console.log(error)
+    } catch (exception:any) {
+      error('Oops!','There was something issue')
     } finally {
       loading.value = false
     }
@@ -46,9 +49,11 @@ export const useTaskStore = defineStore('taskStore', () => {
       await createTask(form.value)
       await fetchTasks()
       router.push({ name: 'task.list' })
-    } catch (error: any) {
-      if (error?.response.status === 422) {
-        errors.value = error.response.data.errors
+      success('Task saved successfully')
+    } catch (exception: any) {
+      if (exception?.response.status === 422) {
+        errors.value = exception.response.data.errors
+        error('Oops!','There was something issue')
       }
     } finally {
       loading.value = false
@@ -63,8 +68,8 @@ export const useTaskStore = defineStore('taskStore', () => {
 
     try {
       form.value = await getTaskById(id)
-    } catch (error: any) {
-      console.log(error)
+    } catch (exception: any) {
+      error('Oops!','There was something issue')
     } finally {
       loading.value = false
     }
@@ -80,26 +85,38 @@ export const useTaskStore = defineStore('taskStore', () => {
       await updateTaskById(form.value)
       await fetchTasks()
       router.push({ name: 'task.list' })
-    } catch (error: any) {
-      if (error?.response.status === 422) {
-        errors.value = error.response.data.errors
+      success('Task updated successfully')
+    } catch (exception: any) {
+      if (exception?.response.status === 422) {
+        errors.value = exception.response.data.errors
+        error('Oops!','There was something issue')
       }
     } finally {
       loading.value = false
     }
   }
   const deleteTask = async (id: number) => {
-    if (loading.value) return
+    confirm({
+      title: 'Are you sure?',
+      text: 'You won\'t be able to revert this action!',
+      icon: 'warning',
+    })
+      .then(async result => {
+        if (result.isConfirmed) {
+          if (loading.value) return
 
-    deletingItem.value = id
-    try {
-      await deleteTaskById(id)
-      await fetchTasks()
-    } catch (error) {
-      console.log(error)
-    } finally {
-      deletingItem.value = 0
-    }
+          deletingItem.value = id
+          try {
+            await deleteTaskById(id)
+            await fetchTasks()
+            success('Task deleted successfully')
+          } catch (exception:any) {
+            error('Oops!','There was something issue')
+          } finally {
+            deletingItem.value = 0
+          }
+        }
+      })
   }
 
   return {
